+++ packages/cards/src/runtime/core/index.js
@@ -2,3 +2,13 @@ export { matchWhen } from './matchWhen.js'
 export { runSteps } from './runSteps.js'
 export { getSteps, normalizeCardTriggers } from './cardTriggers.js'
 export { createCardEngine } from './createCardEngine.js'
+export {
+  evaluateTriggers,
+  triggersNeedAnswer,
+} from './evaluateTriggers.js'
+export {
+  listEffects,
+  attachPrompt,
+  resolveDocEffects,
+  collectMomentSources,
+} from './resolveEffects.js'